bitcoin: move construct_witness from transaction.py to bitcoin.py
